# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AutomaticReform": {
        "additionalProperties": false,
        "properties": {
          "dimple": {
            "default": false,
            "title": "Dimple",
            "type": "boolean"
          },
          "mechanism": {
            "enum": [
              "zipper",
              "string"
            ],
            "title": "Mechanism",
            "type": "string"
          },
          "turn_knot": {
            "default": false,
            "title": "Turn Knot",
            "type": "boolean"
          },
          "wearer_height_cm": {
            "exclusiveMinimum": 0.0,
            "title": "Wearer Height Cm",
            "type": "number"
          }
        },
        "required": [
          "mechanism",
          "wearer_height_cm"
        ],
        "title": "AutomaticReform",
        "type": "object"
      },
...
      "CartItemIn": {
        "properties": {
          "applied_user_coupon_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied User Coupon Id"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "item_type": {
            "enum": [
              "product",
              "reform"
            ],
            "title": "Item Type",
            "type": "string"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "reform_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReformDataIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "selected_option_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option Id"
          }
        },
        "required": [
          "item_id",
          "item_type",
          "quantity"
        ],
        "title": "CartItemIn",
        "type": "object"
      },
      "CartItemOut": {
        "properties": {
          "applied_coupon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserCouponOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "item_type": {
            "title": "Item Type",
            "type": "string"
          },
          "product": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "reform_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReformDataOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "selected_option": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductOptionOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "item_id",
          "item_type",
          "quantity",
          "product",
          "selected_option",
          "reform_data",
          "applied_coupon"
        ],
        "title": "CartItemOut",
        "type": "object"
      },
      "CartRemoveRequest": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Item Ids",
            "type": "array"
          }
        },
        "required": [
          "item_ids"
        ],
        "title": "CartRemoveRequest",
        "type": "object"
      },
      "CartReplaceRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartItemIn"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CartReplaceRequest",
        "type": "object"
      },
...
      "CouponOut": {
        "properties": {
          "additional_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Info"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discount_type": {
            "title": "Discount Type",
            "type": "string"
          },
          "discount_value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Discount Value",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "expiry_date": {
            "format": "date",
            "title": "Expiry Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "max_discount_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Discount Amount"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "display_name",
          "discount_type",
          "discount_value",
          "max_discount_amount",
          "description",
          "expiry_date",
          "additional_info",
          "is_active"
        ],
        "title": "CouponOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductOptionOut": {
        "properties": {
          "additional_price": {
            "title": "Additional Price",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stock": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock"
          }
        },
        "required": [
          "id",
          "name",
          "additional_price",
          "stock"
        ],
        "title": "ProductOptionOut",
        "type": "object"
      },
      "ProductOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detail_images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail Images"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image": {
            "title": "Image",
            "type": "string"
          },
          "info": {
            "title": "Info",
            "type": "string"
          },
          "is_liked": {
            "default": false,
            "title": "Is Liked",
            "type": "boolean"
          },
          "likes": {
            "default": 0,
            "title": "Likes",
            "type": "integer"
          },
          "material": {
            "title": "Material",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "option_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option Label"
          },
          "options": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProductOptionOut"
            },
            "title": "Options",
            "type": "array"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "integer"
          },
          "stock": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "price",
          "image",
          "detail_images",
          "category",
          "color",
          "pattern",
          "material",
          "info",
          "stock",
          "option_label",
          "created_at",
          "updated_at"
        ],
        "title": "ProductOut",
        "type": "object"
      },
...
      "ReformDataIn": {
        "additionalProperties": false,
        "properties": {
          "tie": {
            "$ref": "#/components/schemas/ReformTieIn"
          }
        },
        "required": [
          "tie"
        ],
        "title": "ReformDataIn",
        "type": "object"
      },
      "ReformDataOut": {
        "additionalProperties": false,
        "properties": {
          "cost": {
            "title": "Cost",
            "type": "integer"
          },
          "tie": {
            "$ref": "#/components/schemas/ReformTieOut"
          }
        },
        "required": [
          "tie",
          "cost"
        ],
        "title": "ReformDataOut",
        "type": "object"
      },
      "ReformImageIn": {
        "additionalProperties": false,
        "properties": {
          "claim_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Token"
          },
          "object_key": {
            "minLength": 1,
            "title": "Object Key",
            "type": "string"
          }
        },
        "required": [
          "object_key"
        ],
        "title": "ReformImageIn",
        "type": "object"
      },
      "ReformImageOut": {
        "additionalProperties": false,
        "properties": {
          "object_key": {
            "title": "Object Key",
            "type": "string"
          }
        },
        "required": [
          "object_key"
        ],
        "title": "ReformImageOut",
        "type": "object"
      },
...
      "ReformTieIn": {
        "additionalProperties": false,
        "properties": {
          "automatic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AutomaticReform"
              },
              {
                "type": "null"
              }
            ]
          },
          "image": {
            "$ref": "#/components/schemas/ReformImageIn"
          },
          "restoration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RestorationReform"
              },
              {
                "type": "null"
              }
            ]
          },
          "width": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WidthReform"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "image"
        ],
        "title": "ReformTieIn",
        "type": "object"
      },
      "ReformTieOut": {
        "additionalProperties": false,
        "properties": {
          "automatic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AutomaticReform"
              },
              {
                "type": "null"
              }
            ]
          },
          "image": {
            "$ref": "#/components/schemas/ReformImageOut"
          },
          "restoration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RestorationReform"
              },
              {
                "type": "null"
              }
            ]
          },
          "width": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WidthReform"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "image"
        ],
        "title": "ReformTieOut",
        "type": "object"
      },
...
      "RestorationReform": {
        "additionalProperties": false,
        "properties": {
          "memo": {
            "default": "",
            "maxLength": 200,
            "title": "Memo",
            "type": "string"
          }
        },
        "title": "RestorationReform",
        "type": "object"
      },
...
      "UserCouponOut": {
        "properties": {
          "coupon": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CouponOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "coupon_id": {
            "format": "uuid",
            "title": "Coupon Id",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used At"
          }
        },
        "required": [
          "id",
          "coupon_id",
          "status",
          "issued_at",
          "expires_at",
          "used_at"
        ],
        "title": "UserCouponOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WidthReform": {
        "additionalProperties": false,
        "properties": {
          "target_width_cm": {
            "exclusiveMinimum": 0.0,
            "title": "Target Width Cm",
            "type": "number"
          }
        },
        "required": [
          "target_width_cm"
        ],
        "title": "WidthReform",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "essesion api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cart": {
      "get": {
        "operationId": "get_cart",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CartItemOut"
                  },
                  "title": "Response Get Cart",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Cart",
        "tags": [
          "cart"
        ]
      },
      "put": {
        "operationId": "replace_cart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CartItemOut"
                  },
                  "title": "Response Replace Cart",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace Cart",
        "tags": [
          "cart"
        ]
      }
    },
    "/cart/remove": {
      "post": {
        "operationId": "remove_cart_items",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartRemoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CartItemOut"
                  },
                  "title": "Response Remove Cart Items",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Cart Items",
        "tags": [
          "cart"
        ]
      }
    },
...
  }
}